adaptor restructure, move createDumbComponent to graphql adaptor
