+++ src/adaptors/graphql/components/__tests__/createDumbComponent.spec.js
