+++ src/adaptors/graphql/utils/__tests__/normalize.spec.js
