+++ src/adaptors/graphql/utils/merge.js
