+++ src/adaptors/graphql/utils/normalize.js
