+++ src/index.js
@@ -2,5 +2,5 @@
 
 export Adrenaline from './components/Adrenaline';
 export createContainer from './components/createContainer';
-export createDumbComponent from './components/createDumbComponent';
+export createDumbComponent from './adaptors/graphql/components/createDumbComponent';
 export createGraphQLAdaptor from './adaptors/graphql';
